+++ ft_memset.c
@@ -1,9 +1,9 @@
 #include <libft.h>
 
-void	*ft_memset(void *dest, int c, unsigned int count)
+void	*ft_memset(void *dest, int c, size_t count)
 {
-	unsigned int i;
-	unsigned char *ptr;
+	size_t			i;
+	unsigned char	*ptr;
 
 	ptr = (unsigned char *)dest;
 	i = 0;
