+++ ft_strlen.c
@@ -1,8 +1,8 @@
 #include <libft.h>
 
-unsigned int ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	unsigned int i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
